% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructive.example-package.R
\docType{package}
\name{constructive.example-package}
\alias{constructive.example}
\alias{constructive.example-package}
\title{constructive.example: Example of a package defining a class and methods for 'constructive'}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynkra/constructive.example}
  \item Report bugs at \url{https://github.com/cynkra/constructive.example/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@example.com}

}
\keyword{internal}
