% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructive.example-package.R
\docType{package}
\name{constructive.example-package}
\alias{constructive.example}
\alias{constructive.example-package}
\title{constructive.example: Example of a package defining a class and methods for 'constructive'}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@example.com}

}
\keyword{internal}
