% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructive-qr.R
\name{opts_qr}
\alias{opts_qr}
\title{Constructive options for class 'qr'}
\usage{
opts_qr(constructor = c("qr", "next"), ...)
}
\arguments{
\item{constructor}{String. Name of the function used to construct the object.}

\item{...}{Additional options used by user defined constructors through the \code{opts} object}
}
\value{
An object of class <constructive_options/constructive_options_qr>
}
\description{
These options will be used on objects of class 'qr'.
}
\details{
Depending on \code{constructor}, we construct the object as follows:
\itemize{
\item \code{"qr"} (default): We build the object using \code{qr()}.
\item \code{"next"} : Use the constructor for the next supported class.
}
}
